(#25): 모킹 post 핸들러 추가
