+++ src/entities/assignment/apis/createAssignment.ts
@@ -2,13 +2,14 @@ import type { AssignmentFile } from '@/entities/assignment';
 import { api } from '@/shared';
 
 export const createAssignment = async (file: File) => {
+  const formData = new FormData();
+  formData.append('file', file, file.name);
+
   const response = await api
     .post('assignment', {
-      body: file,
-      headers: {
-        'Content-Type': file.type,
-      },
+      body: formData,
     })
     .json<{ data: AssignmentFile }>();
+
   return response.data;
 };
