+++ src/features/assignment/ui/index.ts
@@ -6,3 +6,4 @@ export { default as AssignmentContainer } from './AssignmentContainer';
 export { default as FileList } from './FileList';
 export { default as FileCard } from './FileCard';
 export { default as AddFileButton } from './AddFileButton';
+export { default as DropZone } from './DropZone';
